# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The REST API a Quark device serves to its Flutter clients. Every endpoint except\n/auth/setup, /auth/login, /auth/recover, /auth/request-account and /auth/status needs a\nsession token. Sign in with POST /auth/login, then click Authorize and enter the word\nBearer, a space, and the token.",
        "title": "Quark API",
        "contact": {},
        "version": "v0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/admin/deny/{username}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a pending account request. Its username is free again at once. Admin-only.",
                "tags": [
                    "admin"
                ],
                "summary": "Deny an account request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username of the pending request",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "no account request has that username",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
        "/admin/groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns every group with its members, the built-in everyone group first and the rest by name. everyone lists no members: it includes every active account. Admin-only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List groups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/grouputil.Group"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates an empty group and its folder in groups on the internal device, named after the group, with one grant giving the group write there; an existing folder of that name is adopted. The name is trimmed, has 1 to 64 characters, no control characters or slashes, isn't . or .., and is unique ignoring case. Publishes account_changed and new_folder. Admin-only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a group",
                "parameters": [
                    {
                        "description": "The group's name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v0_admin.groupNameBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/grouputil.Group"
                        }
                    },
                    "400": {
                        "description": "invalid name",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "409": {
                        "description": "a group with that name already exists",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
        "/admin/groups/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Renames a group under the same rules a new name follows; a group may take its own name in another case. Its folder in groups is renamed to match as an ordinary move does, carrying its shares, favorites and album items, and publishing move and access_changed. The everyone group can't be renamed. The response lists no members. Publishes account_changed. Admin-only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Rename a group",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The group's new name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v0_admin.groupNameBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/grouputil.Group"
                        }
                    },
                    "400": {
                        "description": "invalid name, or the everyone group",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "no group has that id",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "409": {
                        "description": "a group with that name already exists, or another folder in groups already has the new name",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a group, its memberships and every share made to it, so its members lose what only the group gave them. Its folder in groups and its content stay, admin-only from then on. The everyone group can't be deleted. Publishes access_changed with no path. Admin-only.",
                "tags": [
                    "admin"
                ],
                "summary": "Delete a group",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "the everyone group",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "no group has that id",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
        "/admin/users/{username}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an account, its sessions, its shares and its group memberships. Its files stay where they are, and every path it owned becomes owned by the admin deleting it. An admin cannot delete their own account here, and the only active admin cannot be deleted. Admin-only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete an account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username to delete",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v0_admin.deleteUserResponse"
                        }
                    },
                    "400": {
                        "description": "the caller's own account",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "no account has that username",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "409": {
                        "description": "the only active admin",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "grouputil.Group": {
            "type": "object",
            "properties": {
                "builtin": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/grouputil.Member"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "grouputil.Member": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "serverutil.ContentType": {
            "type": "string",
            "enum": [
                "text/html",
                "application/json"
            ],
            "x-enum-varnames": [
                "ContentTypeHTML",
                "ContentTypeJSON"
            ]
        },
        "serverutil.Response": {
            "type": "object",
            "properties": {
                "contentType": {
                    "$ref": "#/definitions/serverutil.ContentType"
                },
                "data": {},
                "error": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "v0_admin.deleteUserResponse": {
            "type": "object",
            "properties": {
                "ownerRowsReassigned": {
                    "type": "integer"
                }
            }
        },
        "v0_admin.groupNameBody": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "A session token from POST /auth/login, entered as the word Bearer, a space, and the token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}